requirements.txt
fastapi
uvicorn
numpy
pandas
requests
boto3
python-dotenv
httpx
pandas